+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - `MenuItem` now receives and uses passed-in rel prop
+  - Also auto appends "noopener noreferrer" to rel if target="\_blank"
 - `Popover` & `usePopover` refactored internally to be consistent with `Dialog` refactor patterns
   - IMPORTANT NOTE: Popover will no longer apply `active` className to target. Instead it now applies `[aria-expanded='true']` instead. `Button*` has been updated to match this change but implementations that depend on the previous behavior will need to be updated.